--- packages/server/src/app/api/dashboard/queue/queue.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { QueueController } from './queue.controller';
-
-describe('Queue Controller', () => {
-    let controller: QueueController;
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            controllers: [QueueController],
-        }).compile();
-
-        controller = module.get<QueueController>(QueueController);
-    });
-
-    it('should be defined', () => {
-        expect(controller).toBeDefined();
-    });
-});
