--- packages/server/src/app/api/dashboard/queue/queue.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { QueueService } from './queue.service';
-
-describe('QueueService', () => {
-    let service: QueueService;
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            providers: [QueueService],
-        }).compile();
-
-        service = module.get<QueueService>(QueueService);
-    });
-
-    it('should be defined', () => {
-        expect(service).toBeDefined();
-    });
-});
