+++ packages/server/src/app/services/admin/index.ts
@@ -1,3 +1,4 @@
 import { MetaService } from './meta.service';
+import { ApprovalQueueService } from './approval-queue.service';
 
-export const AdminServices = [MetaService];
+export const AdminServices = [MetaService, ApprovalQueueService];
